requirements.txt
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy
scipy
pandas
mlflow
scikit-learn
threadpoolctl==2.1.0
Werkzeug==1.0.1
tensorflow-cpu
dvc
pyyaml
optuna
joblib
evidently
pandas-profiling
